requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
Deprecated==1.2.18
django==4.2.23
gunicorn==23.0.0
idna==3.10
importlib-metadata==8.7.0
markdown==3.8.2
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==25.0
prompt-toolkit==3.0.51
pycparser==2.22
PyGithub==2.6.1
pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
python-dotenv==1.1.0
questionary==2.1.0
requests==2.32.4
rich==14.0.0
sqlparse==0.5.3
typing-extensions==4.14.0
urllib3==2.5.0
wcwidth==0.2.13
whitenoise==6.9.0
wrapt==1.17.2
zipp==3.23.0
